Fixes issue #5, not a "perfect solution"
This installs rbenv for both mike & pair, and installs ruby 1.9.3 for both
of them too. We can probably have a better solution with a single global
rbenv (or maybe switch to rvm?)

So the downside is that it takes longer to create the template, but at least
there are users with independent rubies and their gems now.